move request logic to datastores
