+++ packages/runtime/container-runtime/src/containerRuntime.ts
@@ -102,7 +102,6 @@ import {
 	seqFromTree,
 	calculateStats,
 	TelemetryContext,
-	createResponseError,
 } from "@fluidframework/runtime-utils";
 import { GCDataBuilder, trimLeadingAndTrailingSlashes } from "@fluidframework/garbage-collector";
 import { v4 as uuid } from "uuid";
@@ -169,7 +168,6 @@ import {
 	OpSplitter,
 	RemoteMessageProcessor,
 } from "./opLifecycle";
-import { sendGCUnexpectedUsageEvent } from "./garbageCollectionHelpers";
 
 export enum ContainerMessageType {
 	// An op to be delivered to store
@@ -1206,6 +1204,7 @@ export class ContainerRuntime
 			async () => this.garbageCollector.getBaseGCDetails(),
 			(path: string, timestampMs: number, packagePath?: readonly string[]) =>
 				this.garbageCollector.nodeUpdated(path, "Changed", timestampMs, packagePath),
+			(path: string) => this.garbageCollector.isNodeDeleted(path),
 			new Map<string, string>(dataStoreAliasMap),
 		);
 
@@ -1538,27 +1537,6 @@ export class ContainerRuntime
 		}
 		if (typeof request.headers?.[AllowTombstoneRequestHeaderKey] === "boolean") {
 			headerData.allowTombstone = request.headers[AllowTombstoneRequestHeaderKey];
-		}
-
-		const dataStoreNodePath = `/${id}`;
-		if (this.garbageCollector.isNodeDeleted(dataStoreNodePath)) {
-			// The requested data store is removed by gc. Create a 404 gc response exception.
-			const error = responseToException(
-				createResponseError(404, "DataStore was deleted", request),
-				request,
-			);
-			sendGCUnexpectedUsageEvent(
-				this.mc,
-				{
-					eventName: "GC_Deleted_DataStore_Requested",
-					category: "error",
-					isSummarizerClient: this.clientDetails.type === summarizerClientType,
-					headers: JSON.stringify(headerData),
-				},
-				undefined /** packagePath */,
-				error,
-			);
-			throw error;
 		}
 
 		await this.dataStores.waitIfPendingAlias(id);
